New package- utils:
This package will hold togather different utilities we may need.

First module- argument_parsers:
A module to hold different argument parsers for our different main's.

We need a new function to find out the played months (since we can't just assume it's
Aug-May)- I'll work on that tomorrow.

I'll try to change the crawler code to be able to run in parallel (multiprocessing).